luma_demo_construction_company/flask_server/requirements.txt
flask==2.2.3
flask-cors==3.0.10
python-dotenv==1.0.0
requests==2.28.2
jsonschema==4.17.3
gunicorn==20.1.0
psycopg2-binary==2.9.6 